+++ src/uwaterloo/student/zombie/crazy/domain/Sentient.java
@@ -2,5 +2,5 @@ package uwaterloo.student.zombie.crazy.domain;
 
 public interface Sentient {
 	void makeDecision();
-	int getActionTime();
+	long getActionTime();
 }
